Ajustes dos nomes das colunas

+++ src/main/java/br/senac/seguradora/projetoseguradora/controller/ClienteController.java
@@ -38,7 +38,7 @@ public class ClienteController {
 	
 	@PostMapping(path = "/{id}")
 	public Cliente listarPorId(@PathVariable Integer id) {
-		return clienteService.listarPorId(id.longValue());
+		return clienteService.listarPorId(id);
 	}
 	
 	@PostMapping
